+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v2.0.0
+
+`data-scroll-header` -> `data-scroll-top`
+`data-scroll-footer` -> `data-scroll-bottom`
+
+Remove `data-always-show`
 
 ## v1.0.1